REALLY fix the bug where you can't click actors
Have FloorController send click messages to children.

After lots of debugging, the built-in Physics2DRaycaster is fundamentally not hitting the entity, because it looks for a SpriteRenderer *on* the box collider gameobject. But because of animations, I moved the SpriteRenderer onto a child gameobject. So instead of going by individual Entity's Controllers, have the FloorController decide it all.